chore: add jscpd + knip (warn)
