+++ .gitignore
@@ -12,6 +12,9 @@ dist
 dist-ssr
 *.local
 
+# jscpd duplication-scan output
+report
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
