+++ package.json
@@ -32,6 +32,7 @@
     "lint": "eslint src --fix",
     "format": "prettier --write \"src/**/*.{ts,vue,css,html,json}\"",
     "format:check": "prettier --check \"src/**/*.{ts,vue,css,html,json}\"",
+    "knip": "knip",
     "test": "node --import tsx --test test/test_*.ts",
     "test:e2e": "playwright test"
   },
@@ -51,12 +52,12 @@
     "@types/node": "^26.1.2",
     "@vitejs/plugin-vue": "^6.0.8",
     "@vue/tsconfig": "^0.9.1",
-    "entities": "^8.0.0",
     "eslint": "^10.8.0",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-sonarjs": "^4.2.0",
     "eslint-plugin-vue": "^10.10.0",
     "globals": "^17.8.0",
+    "knip": "^6",
     "prettier": "^3.9.6",
     "tailwindcss": "^4.3.3",
     "tsx": "^4.23.1",
